refactor: first draft of registration refactor
