+++ .github/workflows/build.yml
@@ -16,11 +16,12 @@ jobs:
                   node-version: 10
             - run: npm i
             - name: Run linters
-              uses: samuelmeuli/lint-action@v1
+              uses: wearerequired/lint-action@v1
               with:
                   github_token: ${{ secrets.github_token }}
                   eslint: true
                   eslint_dir: ./src
+                  eslint_extensions: js,ts,tsx
                   prettier: true
             - name: Bump version
               if: startsWith(env.BRANCH,'release') == true
