+++ cypress/integration/registration.spec.ts
@@ -3,26 +3,26 @@ import { config } from '../../src/resources/scripts/config';
 const checkForGenericRegistrationElements = () => {
 	cy.get('#loginLogoWrapper').should('exist');
 	cy.get('[data-consultingtype]').should('exist');
-	cy.get('.registration__overline').should('exist');
-	cy.get('.registration__headline').should('exist');
+	cy.get('.registrationForm__overline').should('exist');
+	cy.get('.registrationForm__headline').should('exist');
 	cy.get('#username').should('exist');
 	cy.get('#passwordInput').should('exist');
 	cy.get('#passwordConfirmation').should('exist');
 	cy.get('#dataProtectionCheckbox').should('exist');
 	cy.get('.button__primary').should('exist');
-	cy.get('.registration__toLogin').should('exist');
+	cy.get('.registrationForm__toLogin').should('exist');
 };
 
 describe('registration', () => {
 	describe('addiction', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.suchtberatung.html');
+			cy.visit('/suchtberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.suchtberatung.html');
+			cy.visit('/suchtberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -32,15 +32,15 @@ describe('registration', () => {
 
 	describe('u25', () => {
 		it('should redirect to helpmail when no aid is given', () => {
-			cy.visit('/registration.u25.html');
+			cy.visit('/u25/registration');
 			cy.url().then((url) => {
-				expect(url).to.contain(config.urls.toU25Helpmail);
+				expect(url).to.contain('https://www.u25.de/helpmail/');
 			});
 		});
 
-		it('should have all generic registration page elements', () => {
+		it.only('should have all generic registration page elements', () => {
 			cy.fixture('service.agencies.json').then((agencies) => {
-				cy.intercept(config.endpoints.agencyServiceBase, agencies);
+				cy.visit('/u25/registration?aid=1');
 				cy.visit('/registration.u25.html?aid=1');
 				cy.get('[data-cy=close-welcome-screen]').click();
 				checkForGenericRegistrationElements();
@@ -49,7 +49,7 @@ describe('registration', () => {
 
 		it('should have a password reset info text', () => {
 			cy.fixture('service.agencies.json').then((agencies) => {
-				cy.intercept(config.endpoints.agencyServiceBase, agencies);
+				cy.visit('/u25/registration?aid=1');
 				cy.visit('/registration.u25.html?aid=1');
 				cy.get('[data-cy=close-welcome-screen]').click();
 				cy.get('[data-cy=no-password-reset-possible-note]').should(
@@ -61,7 +61,7 @@ describe('registration', () => {
 		it('should have an agency info when aid is given', () => {
 			cy.fixture('service.agencies.json').then((agencies) => {
 				cy.intercept(config.endpoints.agencyServiceBase, agencies);
-				cy.visit('/registration.u25.html?aid=1');
+				cy.visit('/u25/registration?aid=1');
 				cy.get('[data-cy=close-welcome-screen]').click();
 				cy.get('[data-cy=show-preselected-agency]').should('exist');
 				cy.get('[data-cy=show-preselected-agency]').contains(
@@ -73,13 +73,13 @@ describe('registration', () => {
 
 	describe('pregnancy', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.schwangerschaftsberatung.html');
+			cy.visit('/schwangerschaftsberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.schwangerschaftsberatung.html');
+			cy.visit('/schwangerschaftsberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -89,13 +89,13 @@ describe('registration', () => {
 
 	describe('parenting', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.eltern-familie.html');
+			cy.visit('/eltern-familie/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.eltern-familie.html');
+			cy.visit('/eltern-familie/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -105,13 +105,13 @@ describe('registration', () => {
 
 	describe('cure', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.kurberatung.html');
+			cy.visit('/kurberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.kurberatung.html');
+			cy.visit('/kurberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -121,13 +121,13 @@ describe('registration', () => {
 
 	describe('debt', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.schuldnerberatung.html');
+			cy.visit('/schuldnerberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.schuldnerberatung.html');
+			cy.visit('/schuldnerberatung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -137,13 +137,13 @@ describe('registration', () => {
 
 	describe('social', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.allgemeine-soziale-beratung.html');
+			cy.visit('/allgemeine-soziale-beratung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.allgemeine-soziale-beratung.html');
+			cy.visit('/allgemeine-soziale-beratung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -153,13 +153,13 @@ describe('registration', () => {
 
 	describe('seniority', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.leben-im-alter.html');
+			cy.visit('/leben-im-alter/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.leben-im-alter.html');
+			cy.visit('/leben-im-alter/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -169,17 +169,13 @@ describe('registration', () => {
 
 	describe('disability', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit(
-				'/registration.behinderung-und-psychische-erkrankung.html'
-			);
+			cy.visit('/behinderung-und-psychische-erkrankung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit(
-				'/registration.behinderung-und-psychische-erkrankung.html'
-			);
+			cy.visit('/behinderung-und-psychische-erkrankung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -189,13 +185,13 @@ describe('registration', () => {
 
 	describe('planB', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.mein-planb.html');
+			cy.visit('/mein-planb/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.mein-planb.html');
+			cy.visit('/mein-planb/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -205,13 +201,13 @@ describe('registration', () => {
 
 	describe('law', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.rechtliche-betreuung.html');
+			cy.visit('/rechtliche-betreuung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.rechtliche-betreuung.html');
+			cy.visit('/rechtliche-betreuung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -221,13 +217,13 @@ describe('registration', () => {
 
 	describe('offender', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.straffaelligkeit.html');
+			cy.visit('/straffaelligkeit/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.straffaelligkeit.html');
+			cy.visit('/straffaelligkeit/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -237,13 +233,13 @@ describe('registration', () => {
 
 	describe('aids', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.hiv-aids-beratung.html');
+			cy.visit('/hiv-aids-beratung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.hiv-aids-beratung.html');
+			cy.visit('/hiv-aids-beratung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -253,13 +249,13 @@ describe('registration', () => {
 
 	describe('rehabilitation', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.kinder-reha.html');
+			cy.visit('/kinder-reha/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.kinder-reha.html');
+			cy.visit('/kinder-reha/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -269,13 +265,13 @@ describe('registration', () => {
 
 	describe('children', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.kinder-jugendliche.html');
+			cy.visit('/kinder-jugendliche/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.kinder-jugendliche.html');
+			cy.visit('/kinder-jugendliche/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -285,13 +281,13 @@ describe('registration', () => {
 
 	describe('kreuzbund', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.kb-sucht-selbsthilfe.html');
+			cy.visit('/kb-sucht-selbsthilfe/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.kb-sucht-selbsthilfe.html');
+			cy.visit('/kb-sucht-selbsthilfe/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -301,13 +297,13 @@ describe('registration', () => {
 
 	describe('migration', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.migration.html');
+			cy.visit('/migration/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.migration.html');
+			cy.visit('/migration/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -317,13 +313,13 @@ describe('registration', () => {
 
 	describe('emigration', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.rw-auswanderung.html');
+			cy.visit('/rw-auswanderung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.rw-auswanderung.html');
+			cy.visit('/rw-auswanderung/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -333,13 +329,13 @@ describe('registration', () => {
 
 	describe('hospice', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.hospiz-palliativ.html');
+			cy.visit('/hospiz-palliativ/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.hospiz-palliativ.html');
+			cy.visit('/hospiz-palliativ/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -349,13 +345,13 @@ describe('registration', () => {
 
 	describe('regional', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.regionale-angebote.html');
+			cy.visit('/regionale-angebote/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.regionale-angebote.html');
+			cy.visit('/regionale-angebote/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
@@ -365,13 +361,13 @@ describe('registration', () => {
 
 	describe('men', () => {
 		it('should have all generic registration page elements', () => {
-			cy.visit('/registration.jungen-und-maenner.html');
+			cy.visit('/jungen-und-maenner/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			checkForGenericRegistrationElements();
 		});
 
 		it('should have no password reset info text', () => {
-			cy.visit('/registration.jungen-und-maenner.html');
+			cy.visit('/jungen-und-maenner/registration');
 			cy.get('[data-cy=close-welcome-screen]').click();
 			cy.get('[data-cy=no-password-reset-possible-note]').should(
 				'not.exist'
