+++ package.json
@@ -25,6 +25,7 @@
 		"bfj": "^7.0.2",
 		"camelcase": "^6.1.0",
 		"case-sensitive-paths-webpack-plugin": "2.3.0",
+		"clsx": "^1.1.1",
 		"core-js": "^3.7.0",
 		"css-loader": "4.3.0",
 		"dotenv": "8.2.0",
