+++ src/components/formAccordion/FormAccordion.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { useEffect, useState } from 'react';
 import './formAccordion.styles';
+import { RequiredComponentsInterface } from '../../globalState';
 import { FormAccordionItem } from '../formAccordion/FormAccordionItem';
 import { AgencySelection } from '../agencySelection/AgencySelection';
 import { autoselectPostcodeForConsultingType } from '../agencySelection/agencySelectionHelpers';
@@ -10,17 +11,14 @@ import { RegistrationUsername } from '../registration/RegistrationUsername';
 import { RegistrationAge } from '../registration/RegistrationAge';
 import { RegistrationState } from '../registration/RegistrationState';
 import { RegistrationPassword } from '../registration/RegistrationPassword';
-import {
-	AccordionItemValidity,
-	RequiredComponents
-} from '../registration/registrationHelpers';
+import { AccordionItemValidity } from '../registration/registrationHelpers';
 
 interface FormAccordionProps {
 	consultingType: number;
 	isUsernameAlreadyInUse: boolean;
 	preselectedAgencyData: any;
 	handleFormAccordionData: Function;
-	additionalStepsData?: RequiredComponents;
+	additionalStepsData?: RequiredComponentsInterface;
 }
 
 export const FormAccordion = (props: FormAccordionProps) => {
