--- src/components/registration/initRegistration.ts
@@ -1,3 +1,0 @@
-import { initRegistration } from './Registration';
-
-initRegistration();
