+++ src/globalState/index.ts
@@ -3,6 +3,7 @@ export * from './helpers/stateHelpers';
 export * from './interfaces/AuthDataInterface';
 export * from './interfaces/SessionsDataInterface';
 export * from './interfaces/UserDataInterface';
+export * from './interfaces/ResortDataInterface';
 
 export * from './provider/AcceptedGroupIdProvider';
 export * from './provider/ActiveSessionGroupIdProvider';
