+++ src/globalState/interfaces/ResortDataInterface.ts
@@ -1,0 +1,14 @@
+export type RequiredComponentsInterface = {
+	age?: any;
+	state?: any;
+};
+
+export interface ResortDataInterface {
+	consultingType: number;
+	overline: string;
+	welcomeTitle: string;
+	requiredComponents?: RequiredComponentsInterface;
+	useInformal: boolean;
+	voluntaryComponents?: any[];
+	requiredAidMissingRedirectUrl?: string;
+}
